// apps/inatrace-fe/src/api/model/apiStockOrderEvidenceFieldValue.ts
export interface ApiStockOrderEvidenceFieldValue {
  /**
   * Entity id
   */
  id?: number;
  /**
   * The id of the Processing evidence field from the codebook
   */
  evidenceFieldId?: number;
  /**
   * The field name of the Processing evidence field from the codebook
   */
  evidenceFieldName?: string;
  /**
   * The data type oof the Processing evidence field from the codebook
   */
  evidenceFieldType?: ApiStockOrderEvidenceFieldValue.EvidenceFieldTypeEnum;
  /**
   * Value holder of type String
   */
  stringValue?: string;
  /**
   * Value holder of type Number
   */
  numericValue?: number;
  /**
   * Value holder of type Boolean
   */
  booleanValue?: boolean;
  /**
   * Value holder of type Date
   */
  dateValue?: Date;
}

/**
 * Namespace for property- and property-value-enumerations of ApiStockOrderEvidenceFieldValue.
 */
export namespace ApiStockOrderEvidenceFieldValue {
// ...
  export enum EvidenceFieldTypeEnum {
    STRING = 'STRING',
    TEXT = 'TEXT',
    NUMBER = 'NUMBER',
    INTEGER = 'INTEGER',
    DATE = 'DATE',
    OBJECT = 'OBJECT',
    PRICE = 'PRICE',
    EXCHANGERATE = 'EXCHANGE_RATE',
    TIMESTAMP = 'TIMESTAMP',
  }
// ...
}