// apps/inatrace-fe/src/api/model/apiStockOrderEvidenceTypeValue.ts
export interface ApiStockOrderEvidenceTypeValue {
  /**
   * The id of the Processing evidence type from the codebook
   */
  evidenceTypeId?: number;
  /**
   * The field code of the Processing evidence type from the codebook
   */
  evidenceTypeCode?: string;
  /**
   * The field label of the Processing evidence type from the codebook
   */
  evidenceTypeLabel?: string;
  /**
   * The date of the document upload
   */
  date?: Date;
  document?: ApiDocument;
}

/**
 * Namespace for property- and property-value-enumerations of ApiStockOrderEvidenceTypeValue.
 */
export namespace ApiStockOrderEvidenceTypeValue {
  /**
   * All properties of ApiStockOrderEvidenceTypeValue.
   */
  export enum Properties {
    /**
     * The id of the Processing evidence type from the codebook
     */
    evidenceTypeId = 'evidenceTypeId',
    /**
     * The field code of the Processing evidence type from the codebook
     */
    evidenceTypeCode = 'evidenceTypeCode',
    /**
     * The field label of the Processing evidence type from the codebook
     */
    evidenceTypeLabel = 'evidenceTypeLabel',
    /**
     * The date of the document upload
     */
    date = 'date',
    document = 'document',
  }

  export function formMetadata() {
    return {
      metadata: formMetadata,
      classname: 'ApiStockOrderEvidenceTypeValue',
      vars: [
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'evidenceTypeId',
          classname: 'ApiStockOrderEvidenceTypeValue',
          dataType: 'number',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'evidenceTypeCode',
          classname: 'ApiStockOrderEvidenceTypeValue',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'evidenceTypeLabel',
          classname: 'ApiStockOrderEvidenceTypeValue',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'date',
          classname: 'ApiStockOrderEvidenceTypeValue',
          dataType: 'Date',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          metadata: ApiDocument.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'document',
          classname: 'ApiStockOrderEvidenceTypeValue',
          dataType: 'ApiDocument',
          isPrimitiveType: false,
          isListContainer: false,
          complexType: 'ApiDocument',
        },
      ],
      validators: {
        evidenceTypeId: [],
        evidenceTypeCode: [],
        evidenceTypeLabel: [],
        date: [],
        document: [],
      },
    };
  }

  // export const ApiStockOrderEvidenceTypeValueValidationScheme = {
  //     validators: [],
  //     fields: {
  //               evidenceTypeId: {
  //                   validators: []
  //               },
  //               evidenceTypeCode: {
  //                   validators: []
  //               },
  //               evidenceTypeLabel: {
  //                   validators: []
  //               },
  //               date: {
  //                   validators: []
  //               },
  //               document: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiStockOrderEvidenceTypeValue>;
}