// apps/inatrace-fe/src/api/model/apiActivityProof.ts
/*
 * Copyright(c) 2009 - 2019 Abelium d.o.o.
 * Kajuhova 90, 1000 Ljubljana, Slovenia
 * All rights reserved
 * Copyright(c) 1995 - 2018 T-Systems Multimedia Solutions GmbH
 * Riesaer Str. 5, 01129 Dresden
 * All rights reserved.
 *
 * INATrace Services API
 * INATrace Services API OpenAPI documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the openapi-typescript-angular-generator.
 * https://github.com/alenabelium/openapi-typescript-angular-generator
 * Do not edit the class manually.
 */

import { ApiDocument } from './apiDocument';

/**
 * Activity proofs
 */

export interface ApiActivityProof {
  /**
   * Entity id
   */
  id?: number;
  /**
   * The formal creation date of the document
   */
  formalCreationDate?: Date;
  /**
   * Date until the document is valid
   */
  validUntil?: Date;
  /**
   * The type of the activity proof
   */
  type?: string;
  document?: ApiDocument;
}

/**
 * Namespace for property- and property-value-enumerations of ApiActivityProof.
 */
export namespace ApiActivityProof {
  /**
   * All properties of ApiActivityProof.
   */
  export enum Properties {
    /**
     * Entity id
     */
    id = 'id',
    /**
     * The formal creation date of the document
     */
    formalCreationDate = 'formalCreationDate',
    /**
     * Date until the document is valid
     */
    validUntil = 'validUntil',
    /**
     * The type of the activity proof
     */
    type = 'type',
    document = 'document',
  }

  export function formMetadata() {
    return {
      metadata: formMetadata,
      classname: 'ApiActivityProof',
      vars: [
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'id',
          classname: 'ApiActivityProof',
          dataType: 'number',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'formalCreationDate',
          classname: 'ApiActivityProof',
          dataType: 'Date',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'validUntil',
          classname: 'ApiActivityProof',
          dataType: 'Date',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'type',
          classname: 'ApiActivityProof',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          metadata: ApiDocument.formMetadata,
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'document',
          classname: 'ApiActivityProof',
          dataType: 'ApiDocument',
          isPrimitiveType: false,
          isListContainer: false,
          complexType: 'ApiDocument',
        },
      ],
      validators: {
        id: [],
        formalCreationDate: [],
        validUntil: [],
        type: [],
        document: [],
      },
    };
  }

  // export const ApiActivityProofValidationScheme = {
  //     validators: [],
  //     fields: {
  //               id: {
  //                   validators: []
  //               },
  //               formalCreationDate: {
  //                   validators: []
  //               },
  //               validUntil: {
  //                   validators: []
  //               },
  //               type: {
  //                   validators: []
  //               },
  //               document: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiActivityProof>;
}