// apps/inatrace-fe/src/api/model/apiDocument.ts
/*
 * Copyright(c) 2009 - 2019 Abelium d.o.o.
 * Kajuhova 90, 1000 Ljubljana, Slovenia
 * All rights reserved
 * Copyright(c) 1995 - 2018 T-Systems Multimedia Solutions GmbH
 * Riesaer Str. 5, 01129 Dresden
 * All rights reserved.
 *
 * INATrace Services API
 * INATrace Services API OpenAPI documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the openapi-typescript-angular-generator.
 * https://github.com/alenabelium/openapi-typescript-angular-generator
 * Do not edit the class manually.
 */

/**
 * batch photo
 */

export interface ApiDocument {
  /**
   * Entity id
   */
  id?: number;
  /**
   * storage key (file on system, s3, ...).
   */
  storageKey: string;
  /**
   * document (file) name
   */
  name: string;
  /**
   * content type
   */
  contentType?: string;
  /**
   * size
   */
  size?: number;
}

/**
 * Namespace for property- and property-value-enumerations of ApiDocument.
 */
export namespace ApiDocument {
  /**
   * All properties of ApiDocument.
   */
  export enum Properties {
    /**
     * Entity id
     */
    id = 'id',
    /**
     * storage key (file on system, s3, ...).
     */
    storageKey = 'storageKey',
    /**
     * document (file) name
     */
    name = 'name',
    /**
     * content type
     */
    contentType = 'contentType',
    /**
     * size
     */
    size = 'size',
  }

  export function formMetadata() {
    return {
      metadata: formMetadata,
      classname: 'ApiDocument',
      vars: [
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'id',
          classname: 'ApiDocument',
          dataType: 'number',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: true,
          name: 'storageKey',
          classname: 'ApiDocument',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: true,
          name: 'name',
          classname: 'ApiDocument',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'contentType',
          classname: 'ApiDocument',
          dataType: 'string',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
        {
          isReadOnly: false,
          isEnum: false,
          required: false,
          name: 'size',
          classname: 'ApiDocument',
          dataType: 'number',
          isPrimitiveType: true,
          isListContainer: false,
          complexType: '',
        },
      ],
      validators: {
        id: [],
        storageKey: [['required']],
        name: [['required']],
        contentType: [],
        size: [],
      },
    };
  }

  // export const ApiDocumentValidationScheme = {
  //     validators: [],
  //     fields: {
  //               id: {
  //                   validators: []
  //               },
  //               storageKey: {
  //                   validators: []
  //               },
  //               name: {
  //                   validators: []
  //               },
  //               contentType: {
  //                   validators: []
  //               },
  //               size: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiDocument>;
}